requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
pdfplumber==0.10.2
python-docx==1.0.1
openai==1.3.5
pydantic==2.5.2
python-multipart==0.0.6 